Extract `ComputeSchemaMixin` from `BaseOperator`
This potentially reduces the complexity of using Merlin schemas and the existing schema computation code outside the context of Merlin DAGs.
